refactor: move to terminal directory
